Fix JavaDocs

+++ src/org/joml/Matrix4f.java
@@ -14519,7 +14519,9 @@ public class Matrix4f implements Externalizable, Matrix4fc {
      * @param nearFarDist
      *          the distance between the far and near plane (the near plane will be calculated by this method).
      *          If the special value {@link Float#POSITIVE_INFINITY} is used, the far clipping plane will be at positive infinity.
-     *          If a negative value is used (except for {@link Float#NEGATIVE_INFINITY}) the near and far planes will be swapped.
+     *          If the special value {@link Float#NEGATIVE_INFINITY} is used, the near and far planes will be swapped and 
+     *          the near clipping plane will be at positive infinity.
+     *          If a negative value is used (except for {@link Float#NEGATIVE_INFINITY}) the near and far planes will be swapped
      * @param zeroToOne
      *          whether to use Vulkan's and Direct3D's NDC z range of <tt>[0..+1]</tt> when <code>true</code>
      *          or whether to use OpenGL's NDC z range of <tt>[-1..+1]</tt> when <code>false</code>
@@ -14544,7 +14546,7 @@ public class Matrix4f implements Externalizable, Matrix4fc {
         if (Float.isInfinite(nearFarDist) && nearFarDist < 0.0f) {
             far = near;
             near = Float.POSITIVE_INFINITY;
-        } else if (Double.isInfinite(nearFarDist) && nearFarDist > 0.0f) {
+        } else if (Float.isInfinite(nearFarDist) && nearFarDist > 0.0f) {
             far = Float.POSITIVE_INFINITY;
         } else if (nearFarDist < 0.0f) {
             far = near;
